saufhaengerle/pyproject.toml
[tool.poetry]
name = "saufhaengerle-django-app"
version = "0.1.0"
description = "The Saufhaengerle Django Project"
authors = ["Andreas Huber <andreas.hoenle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-unicorn = "^0.45.0"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-admin-autocomplete-filter = "^0.7.1"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-magic-trailing-comma = "True"
target-version = ['py310']
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | migrations
  | productfeed/config
  | filehandler/config
)/
'''
